requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
idna==2.8
ImageHash==4.0
kiwisolver==1.1.0
lmdb==0.98
matplotlib==3.1.1
numpy==1.17.4
opencv-python==4.1.1.26
pandas==0.25.3
Pillow==6.2.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
requests==2.22.0
scipy==1.4.1
six==1.13.0
urllib3==1.25.7
